# internal/transport/http/apidocs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "企业 AI Coding 能力治理平台。管理接口先登录获取 data.token，再在 Authorize 中填写 AdminBearer（Bearer + 空格 + Token）。Gateway 使用成员 Access Key，与管理员 Token 分开。JSON 中的业务 ID 使用字符串。",
        "title": "Zentrola API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/access-keys/{id}/revoke": {
            "post": {
                "security": [
                    {
                        "AdminBearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "访问密钥"
                ],
                "summary": "撤销 Access Key",
                "parameters": [
                    {
                        "type": "string",
                        "description": "业务 ID（正整数字符串）",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/http.RevokedResponse"
                                        }
                                    }
                                }
                            ]
                        },
                        "headers": {
                            "X-Request-ID": {
                                "type": "string",
                                "description": "请求追踪 ID"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/login": {
            "post": {
                "description": "登录后复制 data.token，在 Authorize 的 AdminBearer 中填写 Bearer + 空格 + Token。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "管理员登录",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/security.LoginResult"
                                        }
                                    }
                                }
                            ]
                        },
                        "headers": {
                            "X-Request-ID": {
                                "type": "string",
                                "description": "请求追踪 ID"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                        }
                    },
                    "429": {
                        "description": "ACCOUNT_LOCKED：账号临时锁定",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/http.LoginLockResponse"
                                        }
                                    }
                                }
                            ]
                        },
                        "headers": {
                            "Retry-After": {
                                "type": "integer",
                                "description": "距离解锁的剩余秒数"
                            },
                            "X-Request-ID": {
                                "type": "string",
                                "description": "请求追踪 ID"
                            }
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "security": [
                    {
                        "AdminBearer": []
                    }
                ],
                "description": "客户端清除 Token；当前实现不在服务端撤销已签发 JWT。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "管理员退出",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/http.LogoutResponse"
                                        }
                                    }
                                }
                            ]
                        },
                        "headers": {
                            "X-Request-ID": {
                                "type": "string",
                                "description": "请求追踪 ID"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/password": {
            "post": {
                "security": [
                    {
                        "AdminBearer": []
                    }
                ],
                "description": "验证当前密码后更新密码，使该账号所有已签发的登录凭证失效；成功后需要重新登录。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "修改当前管理员登录密码",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.ChangePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/http.LogoutResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/http.LoginLockResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/me": {
            "get": {
                "security": [
                    {
                        "AdminBearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "当前管理员",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/admin.Identity"
                                        }
                                    }
                                }
                            ]
                        },
                        "headers": {
                            "X-Request-ID": {
                                "type": "string",
                                "description": "请求追踪 ID"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "admin.Identity": {
            "type": "object",
            "properties": {
                "displayName": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "0"
                },
                "organizationId": {
                    "type": "string",
                    "example": "0"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "github_com_zentrola_zentrola_internal_transport_http.response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "requestId": {
                    "type": "string"
                }
            }
        },
...
        "http.ChangePasswordRequest": {
            "type": "object",
            "required": [
                "currentPassword",
                "newPassword"
            ],
            "properties": {
                "currentPassword": {
                    "type": "string"
                },
                "newPassword": {
                    "description": "NewPassword 至少 6 个字符且不超过 72 个 UTF-8 字节，不能包含空字符，也不能与当前密码相同。",
                    "type": "string"
                }
            }
        },
...
        "http.LoginLockResponse": {
            "type": "object",
            "properties": {
                "lockedUntil": {
                    "type": "string"
                },
                "retryAfterSeconds": {
                    "type": "integer",
                    "example": 900
                }
            }
        },
        "http.LoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "example": "your-admin-password"
                },
                "username": {
                    "type": "string",
                    "example": "admin"
                }
            }
        },
        "http.LogoutResponse": {
            "type": "object",
            "properties": {
                "clearToken": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "http.RevokedResponse": {
            "type": "object",
            "properties": {
                "revoked": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "security.LoginResult": {
            "type": "object",
            "properties": {
                "expiresAt": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "tokenType": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AdminBearer": {
            "description": "管理员登录返回的 Token，填写：Bearer \u003ctoken\u003e。",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "GatewayBearer": {
            "description": "成员 Access Key，填写：Bearer \u003caccess-key\u003e。可用于 OpenAI 或 Anthropic 接口。",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "GatewayKey": {
            "description": "成员 Access Key 原文，仅用于 Anthropic 接口；不要同时发送 Authorization。",
            "type": "apiKey",
            "name": "x-api-key",
            "in": "header"
        }
    }
}